Refactor test suites into packages and handle duplicate inputs in Round class
